backend/user-auth/requirements.txt
bcrypt==4.1.2
PyJWT==2.8.0
psycopg2-binary==2.9.9